requirements2.txt
# Base dependencies that need to be installed first
pydantic==2.6.1; platform_machine == "arm64"
pydantic-core==2.16.2; platform_machine == "arm64"
typing-extensions==4.9.0
pathlib==1.0.1
dataclasses==0.6

# Core dependencies
google-generativeai==0.3.2; platform_machine == "arm64"
openai==1.12.0; platform_machine == "arm64"
urllib3==2.2.0; platform_machine == "arm64"
python-dotenv==1.0.1
requests==2.31.0; platform_machine == "arm64"
websockets==12.0; platform_machine == "arm64"
pandas==2.2.0; platform_machine == "arm64"

# LangChain related
langgraph==0.0.15; platform_machine == "arm64"
langchain-openai==0.0.5; platform_machine == "arm64"
langchain==0.1.5; platform_machine == "arm64"

# Standard library wrappers
argparse==1.4.0
importlib-metadata==7.0.1

# Additional dependencies
protobuf==4.25.2; platform_machine == "arm64"
grpcio==1.60.0; platform_machine == "arm64"
google-api-core==2.15.0; platform_machine == "arm64"